Read config directly from Vault
This breaks the chicken-egg issue that is present. We
manage the vault using the `alcon-api-gateway` project.
However, we also deploy this plugin using the same project. This creates a dependency in the project to
read the vault while it might not exist yet. By reading
directly from the plugin we remove the need for the vault
provider in the Terraform project.